Moved more functions to util
