+++ pythonforandroid/util.py
@@ -1,9 +1,19 @@
 import contextlib
 from os.path import exists
+from os import getcwd, chdir, makedirs
+import io
+import json
+import shutil
+import sys
+from tempfile import mkdtemp
 try:
     from urllib.request import FancyURLopener
 except ImportError:
     from urllib import FancyURLopener
+
+from pythonforandroid.logger import (logger, Err_Fore)
+
+IS_PY3 = sys.version_info[0] >= 3
 
 
 class ChromeDownloader(FancyURLopener):
@@ -42,3 +52,55 @@ def temp_directory():
 def ensure_dir(filename):
     if not exists(filename):
         makedirs(filename)
+
+
+class JsonStore(object):
+    """Replacement of shelve using json, needed for support python 2 and 3.
+    """
+
+    def __init__(self, filename):
+        super(JsonStore, self).__init__()
+        self.filename = filename
+        self.data = {}
+        if exists(filename):
+            try:
+                with io.open(filename, encoding='utf-8') as fd:
+                    self.data = json.load(fd)
+            except ValueError:
+                print("Unable to read the state.db, content will be replaced.")
+
+    def __getitem__(self, key):
+        return self.data[key]
+
+    def __setitem__(self, key, value):
+        self.data[key] = value
+        self.sync()
+
+    def __delitem__(self, key):
+        del self.data[key]
+        self.sync()
+
+    def __contains__(self, item):
+        return item in self.data
+
+    def get(self, item, default=None):
+        return self.data.get(item, default)
+
+    def keys(self):
+        return self.data.keys()
+
+    def remove_all(self, prefix):
+        for key in self.data.keys()[:]:
+            if not key.startswith(prefix):
+                continue
+            del self.data[key]
+        self.sync()
+
+    def sync(self):
+        # http://stackoverflow.com/questions/12309269/write-json-data-to-file-in-python/14870531#14870531
+        if IS_PY3:
+            with open(self.filename, 'w') as fd:
+                json.dump(self.data, fd, ensure_ascii=False)
+        else:
+            with io.open(self.filename, 'w', encoding='utf-8') as fd:
+                fd.write(unicode(json.dumps(self.data, ensure_ascii=False)))
